adding suplementar figures
Supplementary figures (digital version)

Note: Because of printing limitations the ESS figures are available only in the digital version of the thesys. Supplementary figures 17 to 36 can be downloaded at: 
https://github.com/mjanuario/Taxonomy_Januario_Quental_sup_material/tree/master/sup_figs

Supplementary figure 17. ESS per replica for the q_8 parameter (mean preservation for occurrences before 56 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 18. ESS per replica for the q_7 parameter (mean preservation for occurrences between 56 and 33.9 Ma)  in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 19. ESS per replica for the q_6 parameter (mean preservation for occurrences between 33.9 and 23.03 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 20. ESS per replica for the q_5 parameter (mean preservation for occurrences between 23.03 and 15.97 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 21. ESS per replica for the q_4 parameter (mean preservation for occurrences between 15.97 and 11.65 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 22. ESS per replica for the q_3 parameter (mean preservation for occurrences between 11.65 and 5.33 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 23. ESS per replica for the q_2 parameter (mean preservation for occurrences between 5.33 and 2.58 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 24. ESS per replica for the q_1 parameter (mean preservation for occurrences between 2.58 and 1.8 Ma) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 25. ESS per replica for the q_0 parameter (mean preservation for occurrences between 1.8 Ma and the present) in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 26. ESS per replica for the tot_length (total branch length) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 27. ESS per replica for the root_age (age of clade origin) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 28. ESS per replica for the RJ_hp (reversible jump hyper-prior) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 29. ESS per replica for the k_death (number of sampled extinction rate shifts) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 30. ESS per replica for the k_birth (number of sampled speciation/origination rate shifts) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 31. ESS per replica for the alpha (shape parameter of the gamma distributed preservation heterogeneity) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 32. ESS per replica for the q_rate (mean preservation rate for al time series) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 33. ESS per replica for the BD_lik (likelihood of the birth-death) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 34. ESS per replica for the PP_lik (likelihood of the preservation process) parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 35. ESS per replica for the prior parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.

Supplementary figure 36. ESS per replica for the posterior parameter in PyRate. Borders represents preservations models (brown for mG, black for mG+qShift and purple for qShift) and color of boxes represent the dataset (orange for ALL, blue for CAN green for OUR and red for RAW). The dashed line indicates ESS = 200.